Gather expose serialization via com/Extra
