+++ src/com/SerializePartitioning.cpp
@@ -8,8 +8,7 @@
 #include "com/SerializePartitioning.hpp"
 #include "utils/assertion.hpp"
 
-namespace precice::com {
-namespace serialize {
+namespace precice::com::serialize {
 
 SerializedConnectionMap SerializedConnectionMap::serialize(const ConnectionMap &cm)
 {
@@ -241,46 +240,4 @@ SerializedBoundingBoxMap SerializedBoundingBoxMap::broadcastReceive(Communicatio
   sbbm.assertValid();
   return sbbm;
 }
-} // namespace serialize
-
-void sendConnectionMap(Communication &communication, int rankReceiver, const mesh::Mesh::ConnectionMap &cm)
-{
-  serialize::SerializedConnectionMap::serialize(cm).sendTo(communication, rankReceiver);
-}
-
-void receiveConnectionMap(Communication &communication, int rankSender, mesh::Mesh::ConnectionMap &cm)
-{
-  cm = serialize::SerializedConnectionMap::receiveFrom(communication, rankSender).toConnectionMap();
-}
-
-void broadcastSendConnectionMap(Communication &communication, const mesh::Mesh::ConnectionMap &cm)
-{
-  serialize::SerializedConnectionMap::serialize(cm).broadcastSend(communication);
-}
-
-void broadcastReceiveConnectionMap(Communication &communication, mesh::Mesh::ConnectionMap &cm)
-{
-  cm = serialize::SerializedConnectionMap::broadcastReceive(communication).toConnectionMap();
-}
-
-void sendBoundingBoxMap(Communication &communication, int rankReceiver, const mesh::Mesh::BoundingBoxMap &bbm)
-{
-  serialize::SerializedBoundingBoxMap::serialize(bbm).sendTo(communication, rankReceiver);
-}
-
-void receiveBoundingBoxMap(Communication &communication, int rankSender, mesh::Mesh::BoundingBoxMap &bbm)
-{
-  bbm = serialize::SerializedBoundingBoxMap::receiveFrom(communication, rankSender).toBoundingBoxMap();
-}
-
-void broadcastSendBoundingBoxMap(Communication &communication, const mesh::Mesh::BoundingBoxMap &bbm)
-{
-  serialize::SerializedBoundingBoxMap::serialize(bbm).broadcastSend(communication);
-}
-
-void broadcastReceiveBoundingBoxMap(Communication &communication, mesh::Mesh::BoundingBoxMap &bbm)
-{
-  bbm = serialize::SerializedBoundingBoxMap::broadcastReceive(communication).toBoundingBoxMap();
-}
-
-} // namespace precice::com
+} // namespace precice::com::serialize
