+++ src/m2n/PointToPointCommunication.cpp
@@ -14,8 +14,8 @@
 #include "PointToPointCommunication.hpp"
 #include "com/Communication.hpp"
 #include "com/CommunicationFactory.hpp"
+#include "com/Extra.hpp"
 #include "com/Request.hpp"
-#include "com/SerializeMesh.hpp"
 #include "logging/LogMacros.hpp"
 #include "m2n/DistributedCommunication.hpp"
 #include "mesh/Mesh.hpp"
