+++ src/partition/tests/ProvidedPartitionTest.cpp
@@ -6,9 +6,10 @@
 #include <memory>
 #include <string>
 #include <vector>
+
 #include "com/CommunicateBoundingBox.hpp"
 #include "com/Communication.hpp"
-#include "com/SerializePartitioning.hpp"
+#include "com/Extra.hpp"
 #include "com/SharedPointer.hpp"
 #include "m2n/M2N.hpp"
 #include "mapping/Mapping.hpp"
